Change `Button` component in small screens sidebar to `a` tag
Change `Button` component in small screens sidebar to `a` tag to avoid
the an unnecessary nested ripple effect that `Button` component causes
because it is inside `MenuItem` component which has a ripple effect.

Give the `a` tag `classes.unstyledLink` class to overwrite its default styles.

Remove `.sidebarLink` class from `navbar` styles file because the only
component that uses this class has been changed.